+++ InscryptionAPI/Card/AbilityManager.cs
@@ -235,6 +235,19 @@ public static class AbilityManager
             if (Part2ModularAbilities.BasePart2Modular.Contains(ability.ability))
                 ability.SetDefaultPart2Ability();
 
+            if (name == "DeathShield")
+            {
+                ability.SetPassive(false);
+                baseGame.Add(new FullAbility
+                (
+                    null,
+                    ability.ability,
+                    ability,
+                    typeof(APIDeathShield),
+                    useReversePatch ? OriginalLoadAbilityIcon(name) : AbilitiesUtil.LoadAbilityIcon(name)
+                ));
+                continue;
+            }
             baseGame.Add(new FullAbility
             (
                 null,